pikpak-plus-server/requirements.txt
flask[async]==3.0.0
flask-cors==4.0.0
supabase>=2.9.0
python-dotenv==1.0.0
python-dateutil>=2.9.0
requests==2.31.0
diskcache>=5.6.3
gunicorn>=21.2.0
gevent>=23.9.0
Flask-APScheduler==1.13.1
httpx>=0.27.0,<0.28.0
websockets>=13.0
redis>=5.0.0
nest-asyncio>=1.5.8